use long long

+++ 18-bit-manipulation/challenge/bitwise_operators.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int decial_to_binary(int n)
+long long decial_to_binary(int n)
 {
     long long binary_number = 0;
     int remainder;
@@ -24,12 +24,12 @@ int main()
     scanf("%d", &num1);
     scanf("%d", &num2);
 
-    printf("Complements: %d\n", decial_to_binary(~num1));
-    printf("Result of AND operator: %d\n", decial_to_binary(num1 & num2));
-    printf("Result of OR operator: %d\n", decial_to_binary(num1 | num2));
-    printf("Result of XOR operator: %d\n", decial_to_binary(num1 ^ num2));
-    printf("Right shift 2: %d\n", decial_to_binary(num1 >> 2));
-    printf("Left shift 2: %d\n", decial_to_binary(num1 << 2));
+    printf("Complements: %lld\n", decial_to_binary(~num1));
+    printf("Result of AND operator: %lld\n", decial_to_binary(num1 & num2));
+    printf("Result of OR operator: %lld\n", decial_to_binary(num1 | num2));
+    printf("Result of XOR operator: %lld\n", decial_to_binary(num1 ^ num2));
+    printf("Right shift 2: %lld\n", decial_to_binary(num1 >> 2));
+    printf("Left shift 2: %lld\n", decial_to_binary(num1 << 2));
 
     return 0;
 }
